Add mkfs.jffs2 compatible device table support
This patch adds support for mkfs.jffs2 compatible
device table files. I have also included a sample
device table file (the one from the MTD cvs tree) as
well as a patch adding support for them to mkcramfs. I
also fixed a few bugs I found in the cramfs 1.1
release. It would be very nice to see this applied to
the upstream source tree and released. It makes it
easy to build full filesystems (including device nodes,
proper permissions, etc) without needing to be root,
and is therefore used by the buildsystems for a number
of projects.