implement the creation of more block types
This makes for 4 possibilities: either traditional or direct,
and in each case either compressed or uncompressed.

The direct compressed block has no real advantage over the traditional one
except for repositioning the block sequence. The likely scenario is a
sequence of direct non compressed blocks for XIP located out-of-line
with other such blocks so the needed page alignment is padded only once
for the whole image. If only a portion of a file is represented with
direct uncompressed blocks then one direct compressed block would be
needed to bring the stream back to the compressed block area, after
which standard compressed blocks may occur.