auto-XIP of ELF loadable segments
Currently supports no-MMU (-X) and MMU (-X -X). The only difference
at the moment is the block alignment since on no-MMU it is not required
to be page aligned.

To mitigate the large alignment requirement, the MMU case could leverage
the direct block pointer's ability to gather all XIPable blocks together
at the end of the image (TODO).